[phase-01] Fix the Linux, macOS arm64 and MinGW CI builds
Three failures on PR #190, none visible on the macOS dev machine:

- Linux: InstanceList returned totalTimePlayed() -- an int64_t -- as a
  QVariant. int64_t is long on LP64 Linux and QVariant has no
  constructor for long, so the conversion was ambiguous; on macOS it is
  long long and happened to match. It is now returned as qint64.
- macOS arm64: check-plugin-independence.sh used mapfile, which the
  runners' bash 3.2 does not have; the step died with "command not
  found". It collects the targets with a while-read loop instead
  (checked with /bin/bash 3.2: 15 .mmco targets clean).
- MinGW: the MSYS2 environment installed Qt without qt6-declarative, so
  configuring failed on the Qml component the QML shell requires.

The macOS x86_64 job failed installing ccache (a Homebrew checksum
mismatch on the runner), unrelated to this branch.

+++ launcher/InstanceList.cpp
@@ -291,7 +291,9 @@ QVariant InstanceList::data(const QModelIndex& index, int role) const
 			return pdata->lastLaunch();
 		}
 		case TotalTimePlayedRole: {
-			return pdata->totalTimePlayed();
+			// int64_t is long on LP64 Linux, which QVariant has no
+			// constructor for; qint64 (long long) it has.
+			return static_cast<qint64>(pdata->totalTimePlayed());
 		}
 		case GameVersionRole: {
 			return pdata->gameVersion();
